Initial motor integration
This integration is untested.

Added the motor files from the motor side project (as I could not get the project working on my Pi). This includes part of my library for wiringPi located in the new Libraries folder.

Changes to the motor structure now allows for integration with the library and implementation of a PID controller.